+++ CHANGES.rst
@@ -7,6 +7,8 @@ Unreleased (4.2.0) (XXXX-XX-XX)
 - Annotation default not present in portal, if present and active: add and turn
   on.
 
+- Document.title DDSC if ddsc.
+
 
 Release 4.1.15 (2016-10-10)
 ---------------------